+++ src/pages/seasons-rates/seasons-rates.js
@@ -10,6 +10,7 @@ import UI from 'stores/shuri-ui-store';
 import RateCreateModal from 'pages/seasons-rates/rate-create-modal';
 import { getRates, getSeasons, removeRate, createRate, getContractAccommodations, getContract } from 'providers/api';
 import DialogModal from 'parts/dialog-modal';
+import { getRatesTree } from 'utils/ui-utils';
 
 @observer
 class SeasonsRates extends React.Component {
@@ -350,37 +351,6 @@ class SeasonsRates extends React.Component {
             </>
         );
     }
-}
-
-function getRatesTree(ratesList) {
-    const ratesTree = new Map();
-    ratesList.forEach((rate) => {
-        const { seasonId, roomId } = rate;
-        const roomsMap = ratesTree.get(seasonId)?.data;
-        if (!roomsMap) {
-            ratesTree.set(
-                seasonId,
-                {
-                    isExpanded: false,
-                    data: new Map([[roomId, {
-                        isExpanded: false,
-                        data: [rate]
-                    }]])
-                }
-            );
-        } else if (!roomsMap.has(roomId)) {
-            roomsMap.set(
-                roomId,
-                {
-                    isExpanded: false,
-                    data: [rate]
-                }
-            );
-        } else {
-            roomsMap.get(roomId).data.push(rate);
-        }
-    });
-    return ratesTree;
 }
 
 SeasonsRates.propTypes = {
